# public/openapi.yaml
openapi: 3.1.0
info:
  title: Pipelite REST API
  version: 1.0.0
  description: |
    CRM API for organizations, people, deals, activities, and more.
    
    ## Authentication
    
    All API endpoints require Bearer token authentication using an API key.
    API keys can be created from the Pipelite dashboard and must have the `pk_live_` prefix.
    
    Include the API key in the Authorization header:
    ```
    Authorization: Bearer pk_live_your_api_key_here
    ```
    
    ## Rate Limiting
    
    API requests are rate limited per API key. Rate limit headers are included in responses.
    When rate limited, a 429 status is returned with a `Retry-After` header.
    
    ## Pagination
    
    List endpoints support pagination via `offset` and `limit` query parameters.
    Default limit is 50 items. Maximum limit is 100.
    
    Responses include a `meta` object with pagination info:
    ```json
    {
      "data": [...],
      "meta": { "total": 100, "offset": 0, "limit": 50 }
    }
    ```
    
    ## Expanding Relations
    
    Many endpoints support an `expand` parameter to include related entities.
    Multiple relations can be expanded with comma separation: `?expand=owner,organization`
    
    ## Webhooks
    
    Subscribe to events by creating webhook endpoints. Webhooks are signed with HMAC-SHA256.
    Verify signatures using the `X-Webhook-Signature` header.
servers:
  - url: /api/v1
    description: Current server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key with pk_live_ prefix
  schemas:
...
    PaginatedMeta:
      type: object
      required: [total, offset, limit]
      properties:
        total:
          type: integer
          description: Total number of items available
        offset:
          type: integer
          description: Current offset
        limit:
          type: integer
          description: Current limit
...
    Deal:
      type: object
      required: [id, title, stage_id, owner_id, created_at, updated_at]
      properties:
        id:
          type: string
        title:
          type: string
          maxLength: 200
        value:
          type: number
          nullable: true
          description: Deal value in default currency
        stage_id:
          type: string
          description: ID of the current pipeline stage
        organization_id:
          type: string
          nullable: true
        person_id:
          type: string
          nullable: true
        owner_id:
          type: string
        position:
          type: number
          nullable: true
          description: Position within the stage for kanban ordering
        expected_close_date:
          type: string
          format: date-time
          nullable: true
        notes:
          type: string
          nullable: true
        custom_fields:
          type: object
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    DealCreate:
      type: object
      required: [title, stage_id]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        value:
          type: number
          minimum: 0
        stage_id:
          type: string
          minLength: 1
        organization_id:
          type: string
        person_id:
          type: string
        expected_close_date:
          type: string
          format: date-time
        notes:
          type: string
        custom_fields:
          type: object
          description: Custom field values
...
security:
  - BearerAuth: []
paths:
...
  /deals:
    get:
      summary: List deals
      description: Retrieve a paginated list of deals with optional filters
      operationId: listDeals
      tags: [Deals]
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: expand
          in: query
          description: Relations to expand (owner, organization, person, stage)
          schema:
            type: string
        - name: stage_id
          in: query
          description: Filter by stage ID
          schema:
            type: string
        - name: organization_id
          in: query
          description: Filter by organization ID
          schema:
            type: string
        - name: owner_id
          in: query
          description: Filter by owner ID
          schema:
            type: string
      responses:
        '200':
          description: List of deals
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deal'
                  meta:
                    $ref: '#/components/schemas/PaginatedMeta'
    post:
      summary: Create deal
      operationId: createDeal
      tags: [Deals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DealCreate'
      responses:
        '201':
          description: Deal created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Deal'
...